refactor: replace FloatingDevMenu with modular DevMenu component
- Removed legacy `FloatingDevMenu.swift`
- Integrated new `DevMenu` into `NewGameView`
- Replaced hardcoded buttons with reusable `DevMenuButton`
- Renamed dev notification names for clarity and consistency
